capstone-forest_shuffle/requirements.txt
torch
torchvision
scikit-learn
requests
Pillow
filelock
easyocr
rapidfuzz
pandas